Add files via upload
IT-project(transportation system) this the first login page i did, I made the main div and made the username and password section with a table and contains four table rows 2 for (Username / password) and the (Don't have an account? register)(Forgetpassword) section provide it w <a> tag to be linked to the register page(Main page).